Backend/requirements.txt

fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.2.1
python-multipart>=0.0.9
google-generativeai>=0.8.6
langchain>=0.3.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
chromadb>=0.5.0
faiss-cpu>=1.7.4
numpy<2.0.0
langchain-text-splitters>=0.0.1
tiktoken>=0.7.0